requirements.txt
mediapipe==0.8.6.2
numpy==1.21.1
scikit-learn==0.24.2
scipy==1.7.0
sklearn==0.0
tqdm==4.62.0
python-dotenv==0.19.0
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.9.0+cpu
